requirements.txt
numpy
tensorflow-gpu
tensorflow_probability
matplotlib